pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shoptalk-ai"
version = "0.1.0"
description = "ShopTalk AI - Personalized shopping assistant with Google ADK"
requires-python = ">=3.12"

dependencies = [
    # Core ADK (WITH extensions for LiteLlm)
    "google-adk[extensions]==1.25.1",

    # Web API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",

    # ML / Embeddings
    "torch>=2.2.0",
    "open_clip_torch>=2.23.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.19.0",
    "accelerate>=0.29.0",
    "numpy>=1.26.0",
    "mobileclip",

    # WebDataset (training utilities)
    "webdataset>=0.2.86",
    "braceexpand>=0.1.7",

    # Image processing
    "pillow>=10.3.0",

    # Database
    "pgvector>=0.2.4",
    "asyncpg==0.29.0",
    "sqlalchemy>=2.0.29",
    "faiss-cpu>=1.8.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "tqdm>=4.66.2",

    # AWS (SageMaker endpoint calls)
    "boto3>=1.34.0",
]

[tool.setuptools.packages.find]
where = ["."]

